Moved Place and Category to a separate package
