+++ app/src/main/java/ca/appvelopers/mcgillmobile/model/place/Category.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package ca.appvelopers.mcgillmobile.model;
+package ca.appvelopers.mcgillmobile.model.place;
 
 import android.content.Context;
